:truck: Model 파일이동 (#18)
